ci(update workflows): enhanced pull request triggers and scripts
