+++ .github/workflows/reviewer-pull-request.yml
@@ -1,12 +1,15 @@
-name: Add Reviwers Pull Requests
+name: Reviwers Pull Requests
 
 on:
-  schedule:
-    - cron: "0 * * * *"
+  pull_request:
+    types: [review_requested, review_request_removed]
 
 jobs:
   pull-request:
     runs-on: ubuntu-latest
+    env:
+      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      DISCORD_REVIEWS_WEBHOOK_URL: ${{ secrets.DISCORD_REVIEWS_WEBHOOK_URL }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -14,9 +17,10 @@ jobs:
       - name: Install
         uses: ./.github/composite-actions/install
 
-      - name: Update pull requests
-        run: pnpm reviewer:pr
-        env:
-          GITHUB_TOKEN: ${{ secrets.PAT_TOKEN }}
-          DISCORD_REVIEWS_WEBHOOK_URL: ${{ secrets.DISCORD_REVIEWS_WEBHOOK_URL }}
-          DISCORD_HELP_WANTED_WEBHOOK_URL: ${{ secrets.DISCORD_HELP_WANTED_WEBHOOK_URL }}
+      - name: Notification requested reviewer
+        if: ${{ github.event.action == 'review_requested' }}
+        run: pnpm notification:reviewer ${{ github.event.pull_request.number }}
+
+      - name: Notification removed reviewer
+        if: ${{ github.event.action == 'review_request_removed' }}
+        run: pnpm notification:reviewer ${{ github.event.pull_request.number }} --removed=${{ github.event.requested_reviewer.id }}
