+++ package.json
@@ -46,12 +46,12 @@
     "version:next": "changeset version --snapshot next",
     "observe:issue": "tsx scripts/observe-issues",
     "observe:pr": "tsx scripts/observe-pull-requests",
-    "reviewer:pr": "tsx scripts/reviewer-pull-requests",
     "release": "changeset publish",
     "release:dev": "changeset publish --tag dev",
     "release:next": "changeset publish --tag next",
     "report:insight": "tsx scripts/report-insight",
     "notification:release": "tsx scripts/notification-release",
+    "notification:reviewer": "tsx scripts/notification-reviewer",
     "clean": "pnpm -r --parallel exec rimraf dist .turbo *.log node_modules",
     "docs:dev": "pnpm --prefix docs dev",
     "docs:start": "pnpm --prefix docs start",
